add Evt.ChannelId & Evt.FromId
Slack provides ids for channels & users that it expects you to use
for various parts of the UI. Rather than forcing plugins to look
these up as needed, instead provide it as part of the event as a
string. This will also help with Hipchat as a place to provide
channel & user JIDs.