rename [Ff]rom -> [Uu]ser and [Cc]hannel -> [Rr]oom
The from field is used for both source and destination depending
on whether an event is being sent or received. This also makes it
match many places where From is used for User.

Channel was confusing in code that works with both chat channels
and go channels, so it's now room everywhere, which reads a bit
easier.