refactor: remove the generic broker entirely
This went through a couple iterations. At first I wanted to move
the generic broker into its own directory like the other brokers
then establish a forwarding table in hal/router.go. That worked out
but then I noticed that I could use hal.Broker for the type passed
to plugins' Register methods (which can now be an interface! note
for later...). Once I started converting the plugins, it was apparent
that there was no need for the generic broker at all since plugins
could choose to be associated with a broker's type or the interface
and things can be zipped up in the main.go of a given bot.

This removes a special case and makes the event path clearer.

I may yet be able to remove the evt.Brokers since now there's no
forwarding, which will simplify a few other things.