Plugin instances are tied to brokers, not plugins.
Sometimes plugins are specific to a broker, but in practice I've noticed
that most of the plugins are fine across Hipchat/Slack so everybody
shouldn't have to have boilerplate that is only useful to a couple
plugins. Instead, the onus is on authors of broker-specific plugins to
do proper tyupe checking in their code.

Also removes the evt.Brokers and all the related broker stack thing that
turned out to be not worth the complexity it introduced.