Change Prefs to []Pref for literals, add methods.
Changed Prefs from []*Pref to []Pref to make using literals easier. The
savings from using *Pref is minimal since these don't move around a lot
and the convenience of literals in things like Plugin.Settings is worth
it.

Added Clone() to Prefs to make a deep copy of the list. I need to verify
if this approach is still necessary with the change away from pointers.

Added Key() to Prefs to filter a Prefs list by key.