add the ability to SendTable through brokers
The broker decides how it should render the text via AsciiTable,
Utf8Table, and as an image based on either of those (slack).

This allows tables to be displayed consistently in Slack and make them
readable on mobile devices (where even fixed-width text, when it renders
correctly, is clunky to read).