clean up Reply* methods to make way for pref-based routing
I've been putting this off but it's an important feature, so for
now I've organized the code for it to happen and added notes about
what I need to do to implement it. Most of the basics exist in the
Slack broker for the image table replies and that needs to be lifted
over and cleaned up.