simplify SqlInit and make it return error
Also updates callers to check returned errors where appropriate.

The table tracking changes was iffy from the start and most of the
uses of SqlInit so far are already idempotent so it was dead
waiting to be deleted.

Now each user can make the decision whether or not to ignore errors.
Failing to create a table might not be the end of the world for some.