rework console broker and update repl
Made brokers/console's Stdin/Stdout channels public so users
can simply write strings to them. This seems to work well so far.

Moved the stdout code to a utility helper so it can be avoided
e.g. in the repl that implements its own select loop.

Updated examples/repl to use the newly exposed Stdin/Stdout channels
in brokers/console, update the event loop, add a sleep to the
stdin reader to reduce the possibility of races when printing
responses.