add the IsChat field to hal.Evt
This makes it possible for plugins to tell if an event is intended to
be a chat or not. For example, the archive plugin needs to archive chats
but should not archive reaction events or others that are not actual
chats. The reaction tracker can also filter out chats and only do the
typecheck on other events, which is a bit cheaper than checking
everything.