use mysql's utf8mb4 charset
I updated the seppuku plugin and, for giggles, added Japanese aliases
in the regular expressions that it matches, e.g.

!seppuku
!切腹
!zombie
!ゾンビ

Since mysql defaults to the latin1 charset these got mangled by the
database.

One approach might be to add the charset & collation to all of the
tables since I had to edit them anyways (more on that in a sec), but
rather than requiring plugin authors to remember and think about it
hal now defaults to setting the utf8mb4 charset when no other options
follow the final parenthesis in the create table statement. Since this
is a simply appending a string, there's no need to parse the SQL at all
and instead check that it ends with a simple ")" and go for it. If
someone wants to set different options in their plugin, they may do so
and hal will not mangle those statements since they do not end in ")".

All of the sql tables had to be updated for this change since InnoDB
only allows for 761 bytes for index keys. Any varchar(255) had to be
changed to varchar(191). Rather than fixing just that, I normalized all
the text fields to be varchar(191) so they all match now.

Lastly, after connecting to the database, the SqlDB() init function now
specifies the connection charset just to be sure. There is a way to do
it using the mysql driver by setting ?charset= but that requires users
to set their DSN up correctly with those options and it seemed to fiddly
to get right, so here we are.