s/channel/room/ and use empty string defaults
Update to use room instead of channel to align with the rest of hal.

Set defaults for arguments to the empty string to avoid having to
backflips to detect when a user wants defaults from the current
event or not. Instead always default to empty so a !prefs set --key foo
--value bar results in simply "foo/bar" without the
user/room/broker/plugin set.