first pass at configurable reply-via-dm
Touches a lot of code...

The broker interface now includes SendDM. Brokers for services that
don't have it should alias it to Send(). For now, the Hipchat broker
just panics since it /can/ do it but Hipchat took their docs down.

Updated the ~Find~ methods for prefs and added FindKey(key) and
pref.FindKey(key) to make it easy to go from an event to a list of prefs
to One() to automatically get the right settings. This seems to be the
most useful idiom so more changes will likely appear for this pattern.

Added routing to evt.Reply() which is called by Replyf() which makes it
all work based on the "reply-via-dm" pref key.

Updated the pagerduty plugin for the renamed Evt.AsPref() method which
matches convention elsewhere in Hal.