expire cache every 10min, add TODO, shorten dates
Expire the cache every 10 minutes instead of every hour.

Add a TODO to make it configurable via prefs.

Shorten the dates in the oncall output to just dates to make it more
readable. This will probably go away and be replaced by the user's first
contact method once I get to writing that bit of code.