implement automatic caching of pagerduty escalation policies
Uses the new hal.PeriodicFunc functionality to manage the background
worker.

This is a little messier than I like but it's working and makes it so
users don't have to wait the few seconds it takes for !oncall to
download the escalation policies. There are plenty of TODO comments
indicating where cleanups need to happen.