fill in stubs & docs, update test
Fill in the stubbed methods with working code.

Changed the method chaining API so that AddParam returns the Param
allowing additional methods to be called on it, then call Cmd() to
get the command back.

Updated / filled in docs where they were missing/insufficient.

Updated the test for API changes.