move the new Cmd API into hal
I was debating whether or not to keep this separate. It works well as a
stand-alone module right now but I don't like having the separate
import. So instead of requiring lots of imports, it's in the hal
namespace and follows the same pattern as other features of hal - if you
use it, it's ready to go. If you don't want/need it, it's inert and
decoupled just enough that the compiler/linker should be able to
eliminate it (and if not, it's just dead code in the binary).