get keyed parameter parser working as expected
Finally! Positional parameters aren't working yet. Those can happen
tomorrow. This gets keyed (bool flags & key=value & -k v & --key value)
working as expected and passing tests. It required a lot more code
duplication than I expected. The cmdorsubcmd interface was introduced to
break out the conversion from the tmpParamInst type to KVParamInst or
BoolParamInst. This may not be entirely necessary but it seems to be
working and forced me to correct some inconsistencies in a few places,
so it's a net positive for now even if it goes away later on. More
importantly, it keeps the length of the already-too-long Process()
method at a reasonable size.