panic instead of returning nil pointers
Returning nil pointers to a caller has almost no value. If I write a
plugin that creates a situation where a nil can be returned it is almost
certainly a bug and should panic. The other approach to handling this
would be to return an error, but that would defeat the purpose of this
command parser - to significantly reduce the amount of boilerplate required to
create plugins.