add the hal KV store
Add a simple database-backed KV store with ttls that can be used for
things like squelching the autoresponder and stashing values in plugins.

I originally thought about adding somethign to Prefs to allow ephemeral
preferences but it was ugly right away so this is a start towards a k/v stash.