implement topic updater & clean up PD API
In the course of implementing the topic updater in the pagerduty plugin,
I discovered some leftovers from the V1 API implementation that were in
the way that meant changing the signature for the authenticatedGet and
pagedUrl helpers. Since this was already touching all the call sites, I
also removed the no-longer-needed 'domain' parameter. The V1 API took
some parameters as form-encoded while V2 expects them in the URL query.
This meant moving the encoding into pagedUrl to make it consistent. In
any case it looks better this way IMO.