avoid panic if evt.Reply is used outside of the router
The instance field on an event is only set when executing a plugin
instance. When using evt.Reply outside of that context it can cause
panics since evt.instance is nil so check for that and set it to empty
string in evt.AsPref (which is used in evt.Reply) so messages can flow.