add a rudimentary ability to configure plugins on DM
This isn't ideal since it pollutes the core lib with a plugin
(pluginmgr). This seems like a reasonable compromise compared to hooking
into every message to figure out when no plugins are enabled. At this
point, pluginmgr can be called to attach groups.

I might split plugin groups into a separate plugin that can be wired up
instead and then called to configure the plugins on the first message
received from the user. That way only this little chunk of code has to
exist in the router then all the rest can go in the plugin, making it
optional and still providing the functionality folks want.