filter pager aliases by room id to avoid mis-pages
Many of our Slack channels have paging aliases set up that are
duplicate, primarily "us", as in "!page us" or "!pageus". Before this
change, it was possible for !pageus to accidentally pick up the key for
a different room since it was only searching on the key. This now does a
similar search but does a post-query filter to just the room id of the
event so that crossover can't happen.