clean up the autoresponder and make it less spammy
I'm not entirely sure which change got it working reliably, but this
code seems to do what is expected of it. I suspect that the hal KV
caching was getting in the way and that has been stripped off so KV
requests go straight to the database.

Now the autoresponder will squelch itself for 10 minutes per response
and for 2 hours per user.