allow the archive plugins to receive bot-generated messages
The archive table doesn't currently have messages from the bot in it.
This adds the IsBot boolean to the Event structure so that the messages
from the bot itself can be filtered at plugin execution time instead of
in the broker. Also added a BotEvent field to the plugin struct so
plugins can specify if they want to see bot messages or not, defaults to
false.