harden error reporting in hal.Cmd
Changed Cmd.Process() to return an error.

Added mising method SubCmdInst.GetKVParamInst()

Process() now checks for required parameters and returns an error if any
are missing. This led to making the errors more robust:

Updated the error type to use a new interface, NamedParam instead of
interface{}. This seems to be working well and makes the code more
precise.

Updated users of the error type for the new interface.

Added the Name() method to *Param* to implement the interface.