don't spam the room when Google returns 404's
When the calendar id isn't accessible with the bot's crendientials, the
plugin used to reply with the error on every message in the room, which
was quite annoying.

Now it will only do so once an hour, which should be less annoying and
still provide indication that something is wrong.