handle all-day events from Google Calendar (sortof)
the calendar API returns all-day events in a "Date" field instead of the "DateTime" field.
I fixed that, but it doesn't work because Google doesn't populate the TimeZone for Date
even though the docs say it should always be populated on recurring
events. There's no way to set it correctly; all of my prod systems are
on UTC time so the system time won't help. The pref for timezone may not
match the calendar.

I'm checking this in because it *should* work and will work if the
calendar is in UTC.

I'll track down what's going on with the google calendar API some other
time.