add names to IdxParam & fix Get*ParamInst
Positional parameters were clunky to deal with so I added a name field
that is required to create one and added methods to retreive them by
name. This seems to be working well in pluginmgr (coming in next
commit). This required adjusting code in a few places to use the new
field.

While working on that, I discovered that Get*ParamInst was blowing up in
an unexpected way when using e.g. subcmd.GetIdxParamInst(0) with no
arguments. What now happens is that the method checks to see if there is
an possibility of index 0 and returns a zeroed instance with an empty
value and the found/Found() fields set to false.

Updated tests to reflect the new function signatures and added an
assertError to check the return of cmd.Process().