add a Command field to plugins, convert plugins
In the process of creating Slack slash command support it made sense to
add a Command field to plugins so that the router can do an exact match
on events that have a Command field. This way, the hook can convert the
incoming event into a Hal event and fire it into the router so that
plugins can handle things mostly as normal/before.