Add a built-in logger & switch to using it.
The logger built into Hal provides the most common functions of the
stdlib log package. The difference is it can be tapped to get messages
on a channel rather than the standard log package's io.Writer output.

The main purpose is to allow bots to forward logs to a channel without
parsing a byte stream.

In addition, there is a separate Debugf() method that can be disabled &
enabled at runtime. It goes to separate channels and may be tapped (or
not) independently of informational logs.