lay in some groundwork for plugin-enabled reply routing
This is a first pass at the idea of having reply/outbound event routing
work just like inbound does. This adds a few added fields and the stub
necessary for using the out channel that has existed in the router all
along.

More to come.

My internal ticket for this is CORE-829.