ignore empty values just in case, fix GetNodes select
My test database had some garbage from when I was first playing around
with this code. Now that I'm doing more with the data it makes sense to
have the accessors automatically ignore rows with broken identity in the
form of empty values that may occasionally sneak through.

The SELECT in GetNodes was missing a comma before I started. This method
isn't used anywhere yet so no big deal and it's fixed now.