use the 'more' field to control service pagination
This was a bug that was resulting in only the first page of services
being retreived. IIRC when this code was written the 'total' field was
always populated. This seems to have changed for performance reasons
while the v2 API was still in beta, so fair enough. Only the services
code was using total field which was populated with null that
encoding/json seems to have ignored and left total as the empty value of
0.

The other types have been verified to be using the more field instead
and do not need to be updated.