timestamp updates don't seem to be firing - be explicit
At least on Aurora, it isn't happening automatically. The code is
clearly getting called because new hires who show up in the system
appear in the database quickly. Updating the timestamp explicitly makes
more sense in any case, since they may be used to determine if the data
is recent as opposed to being mostly for debugging like many other
timestamps in hal.