strip most special chars & always check teams
Previously, many strings wouldn't match if they had special characters
such as "Foo Bar Team (FBT)" and some folks wouldn't find a match for
"FBT". This has been corrected.

In addition, users of the bot expect team information to be in the
search so it has been changed to always fetch the team list and scan the
list as opposed to only when no other fields matched. This may need
caching added in the future.